Expose useChainInfo (re-use of some info for signer)
